+++ examples/complex/settings.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     azurerm = {
-      version = "=2.40.0"
+      version = "=2.49.0"
     }
     azuread = {
       version = "=1.3.0"