+++ main.tf
@@ -2,7 +2,7 @@ terraform {
   required_version = ">=0.14.4"
   required_providers {
     azurerm = {
-      version = ">=2.40.0"
+      version = ">=2.49.0"
     }
     azuread = {
       version = ">=1.3.0"